Fix services review findings
