+++ src/services/artifacts_tests.rs
@@ -1,4 +1,4 @@
-use super::artifacts::{ArtifactHandle, ArtifactKind};
+use super::artifacts::{ArtifactHandle, ArtifactKind, atomic_write_under};
 use uuid::Uuid;
 
 #[test]
@@ -44,4 +44,47 @@ fn artifact_id_is_stable_for_kind_path_and_hash() {
 
     assert_eq!(one.artifact_id, two.artifact_id);
     assert_eq!(one.kind, ArtifactKind::CrawlManifest);
+}
+
+#[tokio::test]
+async fn atomic_write_under_rejects_parent_traversal() {
+    let temp = tempfile::TempDir::new().expect("tempdir");
+
+    let err = atomic_write_under(temp.path(), "../escape.txt", b"secret")
+        .await
+        .expect_err("parent traversal must be rejected");
+
+    assert!(err.to_string().contains("unsafe artifact relative_path"));
+    assert!(!temp.path().join("../escape.txt").exists());
+}
+
+#[tokio::test]
+async fn atomic_write_under_cleans_temp_file_on_create_failure() {
+    let temp = tempfile::TempDir::new().expect("tempdir");
+    let root_file = temp.path().join("not-a-dir");
+    std::fs::write(&root_file, "occupied").expect("root file");
+
+    let err = atomic_write_under(&root_file, "nested/output.md", b"content")
+        .await
+        .expect_err("root file cannot contain nested artifact");
+
+    assert!(err.to_string().contains("not a directory") || err.to_string().contains("os error"));
+    let temp_files = std::fs::read_dir(temp.path())
+        .expect("read tempdir")
+        .filter_map(Result::ok)
+        .filter(|entry| entry.file_name().to_string_lossy().contains(".tmp-"))
+        .count();
+    assert_eq!(temp_files, 0, "failed writes must not leave temp artifacts");
+}
+
+#[tokio::test]
+async fn atomic_write_under_writes_relative_path_inside_root() {
+    let temp = tempfile::TempDir::new().expect("tempdir");
+
+    let path = atomic_write_under(temp.path(), "screenshots/shot.png", b"png")
+        .await
+        .expect("write artifact");
+
+    assert_eq!(path, temp.path().join("screenshots/shot.png"));
+    assert_eq!(std::fs::read(path).expect("read artifact"), b"png");
 }
