+++ src/services/llm_backend/concurrency.rs
@@ -1,12 +1,14 @@
 use std::error::Error as StdError;
-use std::sync::{Arc, OnceLock};
+use std::sync::{Arc, LazyLock};
 
+use dashmap::DashMap;
 use tokio::sync::{OwnedSemaphorePermit, Semaphore};
 
 #[cfg(test)]
 const DEFAULT_LLM_COMPLETION_CONCURRENCY: usize = 4;
 
-static COMPLETION_SEMAPHORE: OnceLock<Arc<Semaphore>> = OnceLock::new();
+static COMPLETION_SEMAPHORES: LazyLock<DashMap<(String, usize), Arc<Semaphore>>> =
+    LazyLock::new(DashMap::new);
 
 #[cfg(test)]
 fn parse_completion_concurrency_limit(raw: Option<&str>) -> usize {
@@ -19,8 +21,17 @@ fn parse_completion_concurrency_limit(raw: Option<&str>) -> usize {
 pub async fn acquire_completion_permit(
     limit: usize,
 ) -> Result<OwnedSemaphorePermit, Box<dyn StdError + Send + Sync>> {
-    COMPLETION_SEMAPHORE
-        .get_or_init(|| Arc::new(Semaphore::new(limit.clamp(1, Semaphore::MAX_PERMITS))))
+    acquire_completion_permit_for_key("default", limit).await
+}
+
+pub async fn acquire_completion_permit_for_key(
+    key: impl Into<String>,
+    limit: usize,
+) -> Result<OwnedSemaphorePermit, Box<dyn StdError + Send + Sync>> {
+    let normalized_limit = limit.clamp(1, Semaphore::MAX_PERMITS);
+    COMPLETION_SEMAPHORES
+        .entry((key.into(), normalized_limit))
+        .or_insert_with(|| Arc::new(Semaphore::new(normalized_limit)))
         .clone()
         .acquire_owned()
         .await
