+++ src/services/runtime.rs
@@ -38,13 +38,141 @@ pub enum WorkerMode {
 // `dyn ServiceJobRuntime` (object safety). Native async fn in traits (Rust 1.75+)
 // uses RPITIT which makes the trait non-object-safe. Once all callers are
 // converted to generics, this can be removed.
+pub trait ServiceRuntimeIdentity: Send + Sync {
+    fn mode_name(&self) -> &'static str;
+}
+
+pub trait ServiceSqliteRuntimeAccess: Send + Sync {
+    /// Return the runtime's shared SQLite pool when this runtime is backed by
+    /// SQLite. Long-lived surfaces can use this to avoid opening a separate
+    /// pool and re-running migrations for adjacent scheduler/watch operations.
+    fn sqlite_pool(&self) -> Option<Arc<SqlitePool>> {
+        None
+    }
+}
+
+#[async_trait]
+pub trait ServiceJobSubmission: Send + Sync {
+    async fn enqueue(&self, payload: JobPayload) -> BackendResult<Uuid>;
+    async fn enqueue_with_sidecar(
+        &self,
+        payload: JobPayload,
+        sidecar: JobSidecarPayload,
+    ) -> BackendResult<Uuid> {
+        let _ = payload;
+        let _ = sidecar;
+        Err("sidecar enqueue is not supported by this runtime".into())
+    }
+}
+
+#[async_trait]
+pub trait ServiceJobQuery: Send + Sync {
+    async fn wait_for_job(&self, id: Uuid, kind: JobKind) -> BackendResult<String>;
+    async fn job_errors(&self, id: Uuid, kind: JobKind) -> BackendResult<Option<String>>;
+    async fn has_active_jobs(&self, kind: JobKind) -> BackendResult<bool>;
+    async fn notify_worker(&self, kind: JobKind) -> Result<(), Box<dyn Error + Send + Sync>> {
+        let _ = kind;
+        Err("worker notifications are not supported by this runtime".into())
+    }
+    async fn list_jobs(
+        &self,
+        kind: JobKind,
+        limit: i64,
+        offset: i64,
+    ) -> Result<Vec<ServiceJob>, Box<dyn Error + Send + Sync>>;
+    /// List ingest jobs, optionally filtered by source type.
+    ///
+    /// **Default implementation:** fetches up to `limit` rows then post-filters
+    /// in Rust. This is semantically incorrect when `source_filter` is set and
+    /// matching rows number fewer than `limit` — the caller will receive fewer
+    /// rows than expected even if more matching rows exist.
+    ///
+    /// **Override this** in any concrete impl to push the filter into the database.
+    /// If a future impl forgets to override, results will be silently wrong for
+    /// filtered queries on large tables.
+    async fn list_ingest_jobs(
+        &self,
+        source_filter: Option<&str>,
+        limit: i64,
+        offset: i64,
+    ) -> Result<Vec<ServiceJob>, Box<dyn Error + Send + Sync>> {
+        let jobs = self.list_jobs(JobKind::Ingest, limit, offset).await?;
+        if let Some(filter) = source_filter {
+            Ok(jobs
+                .into_iter()
+                .filter(|job| job.source_type.as_deref() == Some(filter))
+                .collect())
+        } else {
+            Ok(jobs)
+        }
+    }
+    async fn job_status(
+        &self,
+        kind: JobKind,
+        id: Uuid,
+    ) -> Result<Option<ServiceJob>, Box<dyn Error + Send + Sync>>;
+
+    /// Count all jobs of a given kind using the shared pool.
+    ///
+    /// Uses the backend's shared SQLite pool directly — avoids calling
+    /// `open_sqlite_pool()` (which re-runs migrations on every call) and avoids
+    /// bypassing `notify()` on enqueue.
+    async fn count_jobs(&self, kind: JobKind) -> Result<i64, Box<dyn Error + Send + Sync>>;
+
+    /// Per-status histogram for a single job kind. Missing statuses are
+    /// absent from the returned map; callers must treat absent as zero.
+    async fn count_jobs_by_status(
+        &self,
+        kind: JobKind,
+    ) -> Result<
+        std::collections::HashMap<crate::jobs::status::JobStatus, i64>,
+        Box<dyn Error + Send + Sync>,
+    >;
+}
+
+#[async_trait]
+pub trait ServiceJobMaintenance: Send + Sync {
+    async fn cancel_job(
+        &self,
+        kind: JobKind,
+        id: Uuid,
+    ) -> Result<bool, Box<dyn Error + Send + Sync>>;
+    async fn cleanup_jobs(&self, kind: JobKind) -> Result<u64, Box<dyn Error + Send + Sync>>;
+    async fn clear_jobs(&self, kind: JobKind) -> Result<u64, Box<dyn Error + Send + Sync>>;
+    async fn recover_jobs(
+        &self,
+        kind: JobKind,
+        stale_threshold_ms: i64,
+    ) -> Result<u64, Box<dyn Error + Send + Sync>>;
+}
+
+#[async_trait]
+pub trait ServiceWorkerControl: Send + Sync {
+    async fn notify_worker(&self, kind: JobKind) -> Result<(), Box<dyn Error + Send + Sync>> {
+        let _ = kind;
+        Err("worker notifications are not supported by this runtime".into())
+    }
+
+    async fn drain_jobs(&self, kind: JobKind) -> Result<WorkerMode, Box<dyn Error + Send + Sync>> {
+        let _ = kind;
+        Ok(WorkerMode::Unsupported(
+            "queue draining is not supported by this runtime",
+        ))
+    }
+
+    async fn start_worker(
+        &self,
+        kind: JobKind,
+    ) -> Result<WorkerMode, Box<dyn Error + Send + Sync>> {
+        self.notify_worker(kind).await?;
+        self.drain_jobs(kind).await
+    }
+}
+
 #[async_trait]
 pub trait ServiceJobRuntime: Send + Sync {
     fn mode_name(&self) -> &'static str;
 
-    /// Return the runtime's shared SQLite pool when this runtime is backed by
-    /// SQLite. Long-lived surfaces can use this to avoid opening a separate
-    /// pool and re-running migrations for adjacent scheduler/watch operations.
     fn sqlite_pool(&self) -> Option<Arc<SqlitePool>> {
         None
     }
@@ -73,16 +201,6 @@ pub trait ServiceJobRuntime: Send + Sync {
         limit: i64,
         offset: i64,
     ) -> Result<Vec<ServiceJob>, Box<dyn Error + Send + Sync>>;
-    /// List ingest jobs, optionally filtered by source type.
-    ///
-    /// **Default implementation:** fetches up to `limit` rows then post-filters
-    /// in Rust. This is semantically incorrect when `source_filter` is set and
-    /// matching rows number fewer than `limit` — the caller will receive fewer
-    /// rows than expected even if more matching rows exist.
-    ///
-    /// **Override this** in any concrete impl to push the filter into the database.
-    /// If a future impl forgets to override, results will be silently wrong for
-    /// filtered queries on large tables.
     async fn list_ingest_jobs(
         &self,
         source_filter: Option<&str>,
@@ -131,15 +249,7 @@ pub trait ServiceJobRuntime: Send + Sync {
         self.drain_jobs(kind).await
     }
 
-    /// Count all jobs of a given kind using the shared pool.
-    ///
-    /// Uses the backend's shared SQLite pool directly — avoids calling
-    /// `open_sqlite_pool()` (which re-runs migrations on every call) and avoids
-    /// bypassing `notify()` on enqueue.
     async fn count_jobs(&self, kind: JobKind) -> Result<i64, Box<dyn Error + Send + Sync>>;
-
-    /// Per-status histogram for a single job kind. Missing statuses are
-    /// absent from the returned map; callers must treat absent as zero.
     async fn count_jobs_by_status(
         &self,
         kind: JobKind,
@@ -172,26 +282,30 @@ pub async fn resolve_runtime_with_workers(
     }
     .map_err(|e| -> Box<dyn Error + Send + Sync> { e.to_string().into() })?;
     Ok(Arc::new(SqliteServiceRuntime {
-        _cfg: cfg,
+        cfg,
         backend: Arc::new(backend),
     }))
 }
 
 pub struct SqliteServiceRuntime {
-    _cfg: Arc<Config>,
+    cfg: Arc<Config>,
     backend: Arc<SqliteJobBackend>,
 }
 
-#[async_trait]
-impl ServiceJobRuntime for SqliteServiceRuntime {
+impl ServiceRuntimeIdentity for SqliteServiceRuntime {
     fn mode_name(&self) -> &'static str {
         "sqlite"
     }
+}
 
+impl ServiceSqliteRuntimeAccess for SqliteServiceRuntime {
     fn sqlite_pool(&self) -> Option<Arc<SqlitePool>> {
         Some(Arc::clone(self.backend.pool()))
     }
+}
 
+#[async_trait]
+impl ServiceJobSubmission for SqliteServiceRuntime {
     async fn enqueue(&self, payload: JobPayload) -> BackendResult<Uuid> {
         self.backend.enqueue(payload).await
     }
@@ -203,7 +317,10 @@ impl ServiceJobRuntime for SqliteServiceRuntime {
     ) -> BackendResult<Uuid> {
         self.backend.enqueue_with_sidecar(payload, sidecar).await
     }
+}
 
+#[async_trait]
+impl ServiceJobQuery for SqliteServiceRuntime {
     async fn wait_for_job(&self, id: Uuid, kind: JobKind) -> BackendResult<String> {
         self.backend.wait_for_job(id, kind).await
     }
@@ -255,6 +372,23 @@ impl ServiceJobRuntime for SqliteServiceRuntime {
         Ok(job_query::service_job(self.backend.pool(), kind, id).await?)
     }
 
+    async fn count_jobs(&self, kind: JobKind) -> Result<i64, Box<dyn Error + Send + Sync>> {
+        Ok(job_query::count_jobs(self.backend.pool(), kind).await?)
+    }
+
+    async fn count_jobs_by_status(
+        &self,
+        kind: JobKind,
+    ) -> Result<
+        std::collections::HashMap<crate::jobs::status::JobStatus, i64>,
+        Box<dyn Error + Send + Sync>,
+    > {
+        Ok(job_query::count_jobs_by_status(self.backend.pool(), kind).await?)
+    }
+}
+
+#[async_trait]
+impl ServiceJobMaintenance for SqliteServiceRuntime {
     async fn cancel_job(
         &self,
         kind: JobKind,
@@ -285,7 +419,10 @@ impl ServiceJobRuntime for SqliteServiceRuntime {
                 .await?,
         )
     }
+}
 
+#[async_trait]
+impl ServiceWorkerControl for SqliteServiceRuntime {
     async fn notify_worker(&self, kind: JobKind) -> Result<(), Box<dyn Error + Send + Sync>> {
         if !self.backend.notify_worker(kind) {
             return Err("no in-process workers running — use `axon serve` or `--wait true`".into());
@@ -294,21 +431,34 @@ impl ServiceJobRuntime for SqliteServiceRuntime {
     }
 
     async fn drain_jobs(&self, kind: JobKind) -> Result<WorkerMode, Box<dyn Error + Send + Sync>> {
-        let pending_at_start = self.count_jobs(kind).await.unwrap_or(0);
-        eprintln!(
-            "draining {} queue ({pending_at_start} pending)...",
-            kind.table_name()
+        let pending_at_start = ServiceJobQuery::count_jobs(self, kind).await.unwrap_or(0);
+        tracing::info!(
+            queue = kind.table_name(),
+            pending_at_start,
+            "draining job queue"
         );
         let started = std::time::Instant::now();
-        let mut secs = 0u64;
+        let timeout = std::time::Duration::from_secs(self.cfg.job_wait_timeout_secs.max(1));
         loop {
-            if !self.has_active_jobs(kind).await? {
+            if !ServiceJobQuery::has_active_jobs(self, kind).await? {
                 break;
             }
+            if started.elapsed() >= timeout {
+                return Err(format!(
+                    "drain_jobs timed out after {}s while draining {} queue",
+                    timeout.as_secs(),
+                    kind.table_name()
+                )
+                .into());
+            }
             tokio::time::sleep(std::time::Duration::from_secs(1)).await;
-            secs += 1;
-            if secs.is_multiple_of(10) {
-                eprintln!("still draining ({secs}s elapsed)...");
+            let elapsed_secs = started.elapsed().as_secs();
+            if elapsed_secs > 0 && elapsed_secs.is_multiple_of(10) {
+                tracing::info!(
+                    queue = kind.table_name(),
+                    elapsed_secs,
+                    "still draining job queue"
+                );
             }
         }
         Ok(WorkerMode::InProcess {
@@ -316,9 +466,109 @@ impl ServiceJobRuntime for SqliteServiceRuntime {
             elapsed_secs: started.elapsed().as_secs(),
         })
     }
+}
+
+#[async_trait]
+impl ServiceJobRuntime for SqliteServiceRuntime {
+    fn mode_name(&self) -> &'static str {
+        ServiceRuntimeIdentity::mode_name(self)
+    }
+
+    fn sqlite_pool(&self) -> Option<Arc<SqlitePool>> {
+        ServiceSqliteRuntimeAccess::sqlite_pool(self)
+    }
+
+    async fn enqueue(&self, payload: JobPayload) -> BackendResult<Uuid> {
+        ServiceJobSubmission::enqueue(self, payload).await
+    }
+
+    async fn enqueue_with_sidecar(
+        &self,
+        payload: JobPayload,
+        sidecar: JobSidecarPayload,
+    ) -> BackendResult<Uuid> {
+        ServiceJobSubmission::enqueue_with_sidecar(self, payload, sidecar).await
+    }
+
+    async fn wait_for_job(&self, id: Uuid, kind: JobKind) -> BackendResult<String> {
+        ServiceJobQuery::wait_for_job(self, id, kind).await
+    }
+
+    async fn job_errors(&self, id: Uuid, kind: JobKind) -> BackendResult<Option<String>> {
+        ServiceJobQuery::job_errors(self, id, kind).await
+    }
+
+    async fn has_active_jobs(&self, kind: JobKind) -> BackendResult<bool> {
+        ServiceJobQuery::has_active_jobs(self, kind).await
+    }
+
+    async fn notify_worker(&self, kind: JobKind) -> Result<(), Box<dyn Error + Send + Sync>> {
+        ServiceWorkerControl::notify_worker(self, kind).await
+    }
+
+    async fn list_jobs(
+        &self,
+        kind: JobKind,
+        limit: i64,
+        offset: i64,
+    ) -> Result<Vec<ServiceJob>, Box<dyn Error + Send + Sync>> {
+        ServiceJobQuery::list_jobs(self, kind, limit, offset).await
+    }
+
+    async fn list_ingest_jobs(
+        &self,
+        source_filter: Option<&str>,
+        limit: i64,
+        offset: i64,
+    ) -> Result<Vec<ServiceJob>, Box<dyn Error + Send + Sync>> {
+        ServiceJobQuery::list_ingest_jobs(self, source_filter, limit, offset).await
+    }
+
+    async fn job_status(
+        &self,
+        kind: JobKind,
+        id: Uuid,
+    ) -> Result<Option<ServiceJob>, Box<dyn Error + Send + Sync>> {
+        ServiceJobQuery::job_status(self, kind, id).await
+    }
+
+    async fn cancel_job(
+        &self,
+        kind: JobKind,
+        id: Uuid,
+    ) -> Result<bool, Box<dyn Error + Send + Sync>> {
+        ServiceJobMaintenance::cancel_job(self, kind, id).await
+    }
+
+    async fn cleanup_jobs(&self, kind: JobKind) -> Result<u64, Box<dyn Error + Send + Sync>> {
+        ServiceJobMaintenance::cleanup_jobs(self, kind).await
+    }
+
+    async fn clear_jobs(&self, kind: JobKind) -> Result<u64, Box<dyn Error + Send + Sync>> {
+        ServiceJobMaintenance::clear_jobs(self, kind).await
+    }
+
+    async fn recover_jobs(
+        &self,
+        kind: JobKind,
+        stale_threshold_ms: i64,
+    ) -> Result<u64, Box<dyn Error + Send + Sync>> {
+        ServiceJobMaintenance::recover_jobs(self, kind, stale_threshold_ms).await
+    }
+
+    async fn drain_jobs(&self, kind: JobKind) -> Result<WorkerMode, Box<dyn Error + Send + Sync>> {
+        ServiceWorkerControl::drain_jobs(self, kind).await
+    }
+
+    async fn start_worker(
+        &self,
+        kind: JobKind,
+    ) -> Result<WorkerMode, Box<dyn Error + Send + Sync>> {
+        ServiceWorkerControl::start_worker(self, kind).await
+    }
 
     async fn count_jobs(&self, kind: JobKind) -> Result<i64, Box<dyn Error + Send + Sync>> {
-        Ok(job_query::count_jobs(self.backend.pool(), kind).await?)
+        ServiceJobQuery::count_jobs(self, kind).await
     }
 
     async fn count_jobs_by_status(
@@ -328,7 +578,7 @@ impl ServiceJobRuntime for SqliteServiceRuntime {
         std::collections::HashMap<crate::jobs::status::JobStatus, i64>,
         Box<dyn Error + Send + Sync>,
     > {
-        Ok(job_query::count_jobs_by_status(self.backend.pool(), kind).await?)
+        ServiceJobQuery::count_jobs_by_status(self, kind).await
     }
 }
