+++ src/services/types/service/options.rs
@@ -1,0 +1,35 @@
+// ── Options / pagination ─────────────────────────────────────────────────────
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct Pagination {
+    pub limit: usize,
+    pub offset: usize,
+}
+
+#[derive(Debug, Clone, PartialEq, Eq)]
+pub struct RetrieveOptions {
+    pub max_points: Option<usize>,
+    pub cursor: Option<String>,
+    pub token_budget: Option<usize>,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum ServiceTimeRange {
+    Day,
+    Week,
+    Month,
+    Year,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct SearchOptions {
+    pub limit: usize,
+    pub offset: usize,
+    pub time_range: Option<ServiceTimeRange>,
+}
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub struct MapOptions {
+    pub limit: usize,
+    pub offset: usize,
+}
